Fixed initisalistion of tile sizes and changed index values to be ints in kernels

+++ kernel_files/macros_cl.cl
@@ -10,12 +10,12 @@
 #endif
 
 #define __kernel_indexes                            \
-    const size_t column = get_global_id(0);			\
-    const size_t row = get_global_id(1);				\
-    const size_t loc_column = get_local_id(0);			\
-    const size_t loc_row = get_local_id(1);			\
-    const size_t lid = loc_row*LOCAL_X + loc_column;	\
-    const size_t gid = row*get_global_size(0) + column;
+    const int column = get_global_id(0);			\
+    const int row = get_global_id(1);				\
+    const int loc_column = get_local_id(0);			\
+    const int loc_row = get_local_id(1);			\
+    const int lid = loc_row*LOCAL_X + loc_column;	\
+    const int gid = row*get_global_size(0) + column;
 
 #define THARR2D(x_offset, y_offset, big_row)        \
     (                                               \
